Introduce qml.sample to sample from an observable (#256)
* Introduce qml.sample to sample from an observable

Issues: #171, #226

* Implement number of samples

The number of samples is attached to the observable via obs.num_samples
at the point when qml.sample is called.

Issues: #171, #226

* Add sample method to default.qubit

The method itself is more or less a reinterpretation of the sampling
performed in expval.

Issues: #171, #226

* Make NotImplementedErrors in var, sample more informative

* Fix typo in the description of var

* Correct sample method to default.qubit and add test

Sampling now works correctly for multi-mode observables and there
is a first test that checks if the shapes of the output arrays
are correct.

Issues: #171, #226

* Simplify the implementation of expval in default.qubit

It will now rely on the sample method. This avoids duplicate
code and will also work for multi-mode observables.

Issues: #171, #226

* Fix test of spectral_decomposition of default.qubit

* Fix signatures of var, sample in default.qubit

* Add a test of sample mean and variance to test_default_qubit

The test samples 2000 Pauli Z expectations from a coherent superposition.
The expectation of this is 0 and the variance is expected to be 1. There
are large deltas in the comparison to ensure that the test does not fail
for statistical reasons.

Issues: #171, #226

* Expand var to calculate approximate variance

At present, var has always calculated analytic variances, which should
only occur if shots=0. Now, it also calculates approximate variances
via sample().var().

* Raise exception if sample is called with n=0 in default.qubit

The default.qubit plugin now throws an error if the method sample is
called with n=0, which will usually happen if the device is in analytic
mode and someone still calls sample with no argument.

Issues: #171, #226

* Raise an exception if differentiation of .sample is attempted

An exception is raised in qnode.jacobian() if there is an observable
with return_type sample in the queue. Furthermore, a test was added to
check the behaviour.

Issues: #171, #226

* Update changelog to reflect introduction of qml.sample

Issues: #171, #226

* Correct test_gradient_exception_on_sample docstring

* Test if the sampled values are the admissible values

Issues: #171, #226

* Throw an error if number of samples is not a positive integer

Issues: #171, #226

* Add sample support to QNode

This commit also includes tests for qml.sample.

Issues: #171, #226

* Remove trailing whitespacs and unidiomatic typecheck

* Fix precision error in test_default_qubit

* Enable combining sample with var and expval in the same circuit

The return types are as follows:
* Single wire circuit
  * sample : ndarray(float)[n_samples]
  * non-sample : float
* Multi-wire circuit
  * sample only
    * same number of shots: ndarray(float)[n_wires, n_samples]
    * different numbers of shots : ndarray(object)[n_wires]
      where each object is an ndarray(float)[n_samples(wire)]
  * sample and non-sample combined: ndarray(object)[n_wires]
    where each object is either a float (non-sample) or an
    ndarray(float)[n_samples] (sample)
  * non-sample : ndarray(float)[n_wires]

Issues: #171, #226

* Fix PEP8 compliance

* Update .github/CHANGELOG.md

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Update pennylane/_device.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Remove print statements from test_measure.py

* Remove superfluous docstrings in default.qubit

* Throw NotImplementedError if sampling of default.gaussian is attempted

The user is recommended to instead install PennyLane-SF. An accompanying
test was added.

* Fix CodeFactor issues

Also reintroduced the docstring for default.qubits' ev method that was
lost in #259

* Fix further CodeFactor issues

* Update pennylane/_device.py

Co-Authored-By: Nathan Killoran <co9olguy@users.noreply.github.com>

* Update pennylane/measure.py

Co-Authored-By: Nathan Killoran <co9olguy@users.noreply.github.com>

* Update pennylane/measure.py

Co-Authored-By: Nathan Killoran <co9olguy@users.noreply.github.com>

* Update pennylane/measure.py

Co-Authored-By: Nathan Killoran <co9olguy@users.noreply.github.com>

* Implement suggestions from code review

* Change self.output_type to self.output_conversion in qnode

* Fix a bug occuring when samples and expvals, vars were returned together

Added an accompanying test.

* Review output conversion in QNode

The code was changed in a way that expects the device to return a
numpy array. In this case the only postprocessing required is to remove
a superfluous array in the case where only a single value is returned.

* Add safety net to QNode type conversion

* Remove statistical test

* Fix erroneous comment

* Increase test coverage

* Test error raised if number of samples not specified for observable

* Fix equal test name

* Test default number of samples = shots

* Test estimation of variance in default.qubit

* Give a more detailed error message if gradients of samples are attempted

* Correct erroneous input type

* Update docstrings and comments in test_var_estimate

* Fix typo

* Fix wrong placing of feature in CHANGELOG.md

* Fix issues turned up by pylint

* Update pennylane/plugins/default_qubit.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Remove trailing whitespaces

* Observable name refactor (#254)

* Add tolerance to test_supported_observables

An assertion of equality was made that failed on Windows due to an
imaginary part ~1e-17. This was replaced with an AlmostEqual assertion
with the default tolerance.

* Change qml.MeanPhoton to qml.NumberOperator

This ensures that the naming of the class is coherent with the new naming
scheme after the introduction of the Observable base class.

Issues: #243

* Change qml.Homodyne to qml.QuadratureOperator

This ensures that the naming of the class is coherent with the new naming
scheme after the introduction of the Observable base class.

Issues: #243

* Replace remaining instaces of Homodyne with QuadratureOperator

Issues: #243

* Change Homodyne to QuadratureOperator in default.gaussian

Issues: #243

* Replace NumberState with NumberStateOperator

This should avoid confusion with the preparation of a number state.

Issues: #243

* Update changelog to reflect renaming of CV observables

* Update .github/CHANGELOG.md

Co-Authored-By: Nathan Killoran <co9olguy@users.noreply.github.com>

* Rename QuadratureOperator to QuadOperator

The naming is more consistent with OpenFermion and StrawberryFields.

Issues: #243

* Rename NumberStateProjector to FockStateProjector

PennyLane tries to exclusively use Fock state if possible.

Issues: #243

* Update .github/CHANGELOG.md

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Fix wrong observable name after merge

* Put changes from observable_name_refactor in correct section in changelog

* Fix typo in quantum neural net example

* Update error message in qml.sample