Add noisy channel: thermal relaxation (#1766)
* Adding a channel ThermalRelaxationError #971

I added the first case of the ThermalRelaxationError channel, it is still needed the second case when T_2 > T_1 but at this point I need help!

* Adding the Choi-Matrix and the Kraus representation for T2 > T1

Adding documentation of the two cases when T2 <= T1 and T2 > T1. It's left to make the test of this noisy model.

* Adding test and improving the Kraus operator for T2 > T1

I add a test for the ThermaRelaxationError and improve the Kraus operators for T2 > T1. Now, I don't use the Choi matrix and extract the eigenvalues and eigenvectors, instead, I use the parametric representation of those eigenvalues and eigenvectors.

* Solving some problems with the test

Solving some of the Issues with the test, now the only problem is that the ThermalRelaxationError is not recognize as a gate.

* Solving issues with the tests

I solve the problems with the tests that were causing errors.

* Adding some test

Adding tests for the errors when the times t1, t2, and tg are negative.

* Trying to fix the documentation problem

* Update channel.py

* Update channel.py

* Update channel.py

* Apply suggestions from code review

Thank you! @rmoyard

Co-authored-by: Romain <rmoyard@gmail.com>

* Adding suggestions

Adding the suggestions of @rmoyard about the reshaping of the eigvecs and adding a test for t2 > t1 to the TestChannels. Additionally, adding the new ThermalRelaxationError to changelog-dev

* Apply suggestions from code review

Co-authored-by: Romain <rmoyard@gmail.com>

* Apply suggestions from code review

Thank you @rmoyard

Co-authored-by: Romain <rmoyard@gmail.com>

* Update changelog-dev.md

Add name

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: antalszava <antalszava@gmail.com>
Co-authored-by: Romain <rmoyard@gmail.com>