Add a parameter-shift Hessian transformation (#1884)
* Initial attempt at a parameter-shift hessian

- includes some basic test cases
- imports 'param_shift_hessian' into the pennylane/gradients module
- currently only computes the diagonal elements of the hessian

* Compute offdiagonal elements of the Hessian

- all elements of the Hessian are now computed
- shape of the returned array is wrong

* Fix shape of returned Hessian

Previously all the elements of the Hessian would be returned in a 1D array
(assuming scalar-valued function, +1D for vector valued functions), a
hold-over from computing gradients.
Now the proper dimensions are returned, with extrenous ones being
stripped (for example for single variable, scalar-valued functions).

* Add test that reuses qnode parameters

This test fails. The added classical processing between the QNode and
the gate parameters reveal that computing the "hybrid" hessian requires
steps beyond those provided by `gradient_transform`.

* Add new gradient transform decorator for Hessian

When generating derivatives of QNodes with both classical and quantum
processing of the QNode inputs, the `gradient_transform` decorator
automatically computes a "hybrid" derivate from the classical jacobian
and quantum jacobian. This only works with first derivates however.
The new decorator behaves similarly while accounting for the second
derivative nature of the hessian.

* Use different recipes for diag/off-diag elements

- Differientiate between diagonal (2 tapes) and off-diagonal (4 tapes)
  elements of the Hessian with different recipes.
- Formalize recipe structure that allows for multiple shifts per tape.
- Adapt `_process_gradient_recipe` for new recipe structure.

* Fix hessian for arbitrary QNode output dimensions

Previously, only QNode output dimensions of up to 1 were considered
(scalars or vectors).

* Support higher dim. cl. Jacobians in hybrid QNodes

Previously, the Hessian transform could only support 2D classical Jacobians
(corresponding to classical processing of QNode parameters).

* Formatting

* Set performance goals for Hessian implementation

Adds new tests to check the number of quantum device invocations is
below a certain target. 3 different targets for 2-term parameter shift
rules are provided.

* Reduce number of gradient tapes generated

Exploits the symmetry of the derivative tensor under index permutation
to generate half the number of tapes for the off-diagonal elements of
the Hessian.

* Fix typo in performance test

Also formatting and linting. Caved to black reformatting my beautiful
matrices.

* Modify Hessian tests for the new QNode

The new QNode requires a new argument `max_diff` to set the desired
differentiation order.
Also includes a few additional assertions that (inicidentally) uncovered
a bug in the new QNode.

* Add 2 optimizations to reduce the # of evaluations

- Reduce the number of gradient tapes generated by only computing the
  unshifted tape (all shifts = 0) once and reusing the result.
- Reduce the number of gradient tapes generated by skipping partial
  derivatives which are known to be zero.

* Fix pylint and std math library dependance

`math.comb` is only supported on Python 3.8 and up so was removed from
the tests.
2 pylint warnings that are failing codefactor are fixed as well.

* Rewrite transformation to avoid array assignments

Array assingments are unsupported in Autograd, so in order to support
computing derivates of the Hessian transform, they need to be removed
from its implementation.
This change also adds a new test to check the correctness of the third
derivate obtained via the Hessian transform.

* Add checks for unsupported operations + tests

* Make Hessian transform framework agnostic

Instead of creating Numpy arrays, use `qml.math.stack` to convert
accumulated  gradients into (flat) arrays/tensors, and reshape to the
proper dimensions at the end.

* Add test for circuit without trainable parameters

* Refactor Hessian transform, increase test coverage

- new input checks and computation shortcuts
- new f0 parameter to unshifted circuit results
- split off the core functionality from preprocessing tasks
- proper docstrings
- new tests for increased code coverage

* More descriptive names, remove print statements

* Review: change Jacobian for Hessian

* Add proper docstring to hessian_transform

* Review: add new functions to docs

* Implement fix for multiple QNode arguments

#1989 introduced a fix to the gradient transform for QNodes with
multiple arguments using autograd. When arguments are of the same shape,
the classical Jacobians are stacked and transposed, leading to a
different tensor contraction with the quantum Jacobian/Hessian than in
the single argument case.

* Add support for tuples of classical Jacobians

The computation is the same as for the single QNode argument case but
repeated with each Jacobian per QNode argument.

* Move hessian_transform to its own file

* Add preliminary tests for all interfaces (failing)

* Review: rename gradient->Hessian

Co-authored-by: David Wierichs <davidwierichs@gmail.com>

* More renaming

* Review: remove redundant dim swaps

* Review: optimize upper triangular optimization

* Review: optimize 0 derivative optimization

* Review: simplify axes reordering

Co-authored-by: David Wierichs <davidwierichs@gmail.com>

* Review: small fixes to the dimension reordering

* Review: clarify wording for 'f0' arg

* Review: prevent silent failure

... when computing the trainable indices across frameworks.

* Fix Hessian transform for Tensorflow

Register the transpose function with autoray to convert the `axes`
keyword argument to `perm`.
Amend test to account for return type being a tuple with tensorflow,
even for single argument QNodes.

* Fix Hessian transform for JAX

Amend test to use backprop, as derivatives of order  > 1 are unsupported
by the JAX interface with parameter-shift. Additionally, change the
QNode return value to expval, as probabilities are unsupported by the
JAX interface.
Stack the tape results before determining their shape. The numpy shape
function converts its argument to ndarray first if it is not already a
single array (in this case, the tape results are a list of arrays). With
JAX, this conversion fails as the list of traced arrays cannot be
converted to a regular ndarray.

* Fix Hessian transform for PyTorch

Similar to JAX, a list of tensors needs to be stacked before
qml.math.shape can be used (fixed in 35a724b).
Amend test to account for return type being a tuple with Torch, even for
single argument QNodes. Remove requires_grad from the tensor passed to
the QNode.

* Mark slow tests

* Review: fix hardcoded expansion function

* Formatting

* Changelog + docstring examples

* Fix coverage with more interface tests

* Formatting

* Changelog improvements

* Revert unrelated file

* Fix docstring examples + add tape drawing

* Add warning and use case to docstring

* Remove stack&transpose workaround

* Adapt multiple argument tests

New changes in #2059 removed the ability to perform higher-order
differentiation in autograd on QNodes with multiple trainable argumemts.
Expected results for these tests are now calculated using JAX instead.

* Improve docstring

Co-authored-by: David Wierichs <davidwierichs@gmail.com>
Co-authored-by: antalszava <antalszava@gmail.com>